refactor: split crawl-courses.ts into modules
