+++ scripts/crawl/client.ts
@@ -1,0 +1,15 @@
+// Service-role Supabase client. Module-load env validation, so importing
+// this module in the entry script is enough to fail fast on missing keys.
+import 'dotenv/config'
+import { createClient } from '@supabase/supabase-js'
+
+const URL = process.env.SUPABASE_URL ?? 'http://127.0.0.1:54321'
+const SERVICE_KEY = process.env.SUPABASE_SERVICE_ROLE_KEY
+if (!SERVICE_KEY) {
+  console.error('SUPABASE_SERVICE_ROLE_KEY is required')
+  process.exit(1)
+}
+
+export const supabase = createClient(URL, SERVICE_KEY, {
+  auth: { autoRefreshToken: false, persistSession: false },
+})
